apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
google-generativeai==0.4.0
pdfplumber==0.10.4
python-multipart==0.0.9
